[RF-2124] Refactor demo frontend to flat structure.
This matches our standard for flat structure in our web app as well, and maintains consistency. This app had some flat components and some non-flat, so this unifies the style throughout to the company style for Angular apps.

Additionally it removes some unhelpful-named directories like "shared", which don't convey any semantic meaning.
