+++ client/app/header.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { StorageService } from '../storage.service';
+import { StorageService } from './storage.service';
 
 /**
  * This component provides a styled header for the application.
