+++ client/app/pay-loans.component.html
@@ -1,4 +1,4 @@
-<img class="logo-title" src="assets/images/logo.svg" alt="logo"/>
+<img class="logo-title" src="../assets/images/logo.svg" alt="logo"/>
 
 <mat-card class="form-card mat-typography">
   <app-header></app-header>
