+++ client/app/pay-loans.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { StorageService } from '../storage.service';
+import { StorageService } from './storage.service';
 import { map, tap } from 'rxjs/operators';
 
 /**
