+++ client/app/payment-form.component.ts
@@ -5,12 +5,12 @@ import { combineLatest, Observable, of, throwError } from 'rxjs';
 import { map, switchMap, take } from 'rxjs/operators';
 import { Payment } from 'rightfoot-node/1-3/api';
 
-import { UnitedStatesTerritories } from '../shared/us-states-and-territories';
-import { phoneNumberValidator } from '../shared/phone-number.validator';
-import { RightfootApiService } from '../rightfoot-api.service';
-import { DemographicsFormValue } from '../shared/demographics-form-value.interface';
-import { StorageService } from '../storage.service';
-import { PlaidService } from '../plaid.service';
+import { UnitedStatesTerritories } from './us-states-and-territories';
+import { phoneNumberValidator } from './phone-number.validator';
+import { RightfootApiService } from './rightfoot-api.service';
+import { DemographicsFormValue } from './demographics-form-value.interface';
+import { StorageService } from './storage.service';
+import { PlaidService } from './plaid.service';
 
 /**
  * This component provides a form with a payment amount input and a "Pay" button.
