+++ client/app/payment-status.component.scss
