+++ client/app/payment-success.component.scss
