+++ client/app/stepper.component.scss
